requirement.txt
Django==1.9
django-ads-txt==0.1.2
django-appconf==1.0.2
django-imagekit==4.0.2
django-markdown-deux==1.0.5
django-pagedown==1.0.5
django-robots==3.1.0
gunicorn==19.9.0
markdown2==2.3.5
pilkit==2.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
six==1.11.0
